requirements.txt
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
asgiref==3.6.0
pytz==2022.7.1
setuptools==65.5.0
sqlparse==0.4.3